# docs/openapi/innovation-sandbox-api.yaml
openapi: 3.0.0
info:
  title: Innovation Sandbox on AWS
  description: API documentation for the Innovation Sandbox on AWS solution.
  version: 1.3.0
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "{apiBaseUrl}"
    description: Base URL used to make API requests
    variables:
      apiBaseUrl:
        default: https://example.com/api
        description: Base URL used to make API requests
...
security:
  - sigV4: []
    isbIdentity: [] # human callers (Identity Pool roles)
  - sigV4: [] # M2M callers (per-client assumed roles)
components:
  securitySchemes:
    sigV4:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: awsSigv4
      description: >-
        AWS Signature Version 4 over the `execute-api` service. Required on
        every request: the API uses IAM authorization
        (`AuthorizationType.IAM`), and there is no bearer-token or Cognito
        authorizer. The caller type is determined from the IAM principal.
    isbIdentity:
      type: apiKey
      name: x-isb-identity
      in: header
      description: >-
        Cognito ID token, verified via JWKS. Carries the claims used for RBAC.
        Required for human callers (Identity Pool roles). M2M callers (per-client
        assumed IAM roles) derive their role from the assumed-role ARN and must
        NOT send this header; a request presenting both an M2M principal and
        `x-isb-identity` is rejected with 400.
  schemas:
    ResponseBody:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - success
            - fail
            - error
      discriminator:
        propertyName: status
        mapping:
          success: "#/components/schemas/SuccessResponseBody"
          fail: "#/components/schemas/FailResponseBody"
          error: "#/components/schemas/ErrorResponseBody"
    SuccessResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - status
            - data
          properties:
            status:
              type: string
              enum:
                - success
            data:
              type: object
    FailResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - data
          properties:
            status:
              type: string
              enum:
                - fail
            data:
              type: object
    ErrorResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - status
            - message
          properties:
            status:
              type: string
              enum:
                - error
            message:
              type: string
            code:
              type: number
            data:
              type: object
    PaginatedResults:
      type: object
      required:
        - result
      properties:
        result:
          type: array
        nextPageIdentifier:
          type: string
          example: "eyJsYXN0RXZhbHVhdGVkS2V5Ijp7InV1aWQiOnsicyI6IjEyMzQ1Njc4LTkwYWItY2RlZi0xMjM0LTU2Nzg5MGFiY2RlZiJ9fX0="
...
    BudgetThresholds:
      type: array
      items:
        type: object
        properties:
          dollarsSpent:
            type: number
            description: Threshold for budget notification
            example: 80
          action:
            type: string
            description: Action to take when threshold is reached
            enum: [ALERT, FREEZE_ACCOUNT]
            example: ALERT
    DurationThresholds:
      type: array
      items:
        type: object
        properties:
          hoursRemaining:
            type: number
            description: Threshold for durations notification
            example: 12
          action:
            type: string
            description: Action to take when threshold is reached
            enum: [ALERT, FREEZE_ACCOUNT]
            example: ALERT
...
    UserEmail:
      type: string
      format: email
      description: Email of the user who owns the lease
      example: "user@example.com"
...
    LeaseTemplateUuid:
      type: string
      format: uuid
      description: Unique identifier for the lease template used to create the lease
      example: "12345678-90ab-cdef-1234-567890abcdef"
    LeaseTemplateName:
      type: string
      description: Name of the lease template
      example: Example Template
    LeaseTemplateDescription:
      type: string
      description: Description of the lease template
      example: Example Template Description
    LeaseDurationInHours:
      type: integer
      description: Duration of the lease in hours
      example: 24
    MaxSpend:
      type: number
      description: Maximum spend allowed for the lease
      example: 100
...
    RequiresApproval:
      type: boolean
      description: Whether leases created from this template require approval
      example: true
...
    Visibility:
      type: string
      enum: [PUBLIC, PRIVATE]
      description: Visibility of the lease template - PUBLIC templates are visible to all users, PRIVATE templates are only visible to Admin and Manager roles
      example: "PUBLIC"
    CostReportGroup:
      type: string
      description: Cost report group identifier for cost allocation and reporting. Must be one of the values configured in costReportGroups if requireCostReportGroup is enabled.
      example: "Engineering"
      maxLength: 50
      nullable: true
    BlueprintUuid:
      type: string
      format: uuid
      description: Unique identifier for the blueprint
      example: "12345678-90ab-cdef-1234-567890abcdef"
...
    MetaData:
      type: object
      properties:
        createdTime:
          type: string
          format: date-time
          description: Time when the resource was created
          example: "2023-05-01T12:00:00Z"
        lastEditTime:
          type: string
          format: date-time
          description: Time when the resource was last edited
          example: "2023-05-01T12:00:00Z"
        schemaVersion:
          type: integer
          description: Version of the schema used to create the resource
          example: 1
...
    LeaseTemplate:
      type: object
      required:
        - uuid
        - name
        - requiresApproval
        - createdBy
      properties:
        uuid:
          $ref: "#/components/schemas/LeaseTemplateUuid"
        name:
          $ref: "#/components/schemas/LeaseTemplateName"
        description:
          $ref: "#/components/schemas/LeaseTemplateDescription"
        requiresApproval:
          $ref: "#/components/schemas/RequiresApproval"
        createdBy:
          $ref: "#/components/schemas/UserEmail"
        visibility:
          $ref: "#/components/schemas/Visibility"
        maxSpend:
          $ref: "#/components/schemas/MaxSpend"
        budgetThresholds:
          $ref: "#/components/schemas/BudgetThresholds"
        leaseDurationInHours:
          $ref: "#/components/schemas/LeaseDurationInHours"
        durationThresholds:
          $ref: "#/components/schemas/DurationThresholds"
        costReportGroup:
          $ref: "#/components/schemas/CostReportGroup"
        blueprintId:
          allOf:
            - $ref: "#/components/schemas/BlueprintUuid"
            - nullable: true
              description: Optional blueprint attached to this lease template
        blueprintName:
          type: string
          nullable: true
          description: Resolved name of the attached blueprint
        allowOwnerToShareLease:
          type: boolean
          description: Whether lease owners created from this template can manage additional user assignments
          example: false
        meta:
          $ref: "#/components/schemas/MetaData"
...
  parameters:
    PageIdentifier:
      name: pageIdentifier
      in: query
      required: false
      schema:
        type: string
      description: Base64 encoded token for pagination
    MaxResults:
      name: maxResults
      in: query
      required: false
      schema:
        type: integer
      description: >-
        Maximum number of items to return per page. The legacy `pageSize`
        parameter is unsupported: most endpoints ignore it and apply their
        default, while `/leases/shared` rejects it as an unknown parameter.
...
    LeaseTemplateIdPathParam:
      name: leaseTemplateName
      description: >-
        Lease template UUID. The path segment is named `leaseTemplateName` in the
        API Gateway definition, but the value is the template's UUID, not its
        display name.
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/LeaseTemplateUuid"
...
  responses:
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors: [{ "message": "example error message" }]
    UnauthenticatedResponse:
      description: >-
        Unauthenticated. The `x-isb-identity` ID token is missing, expired,
        invalid, or lacks required claims.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors:
                [
                  {
                    "message": "Token is missing required claims. Please re-authenticate.",
                  },
                ]
    UnauthorizedResponse:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            message: Unauthorized
    ResourceNotFoundResponse:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors: [{ "message": "example error message" }]
...
    ErrorResponse:
      description: Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseBody"
          example:
            status: error
            message: "An unexpected error occurred"
paths:
...
  /leaseTemplates:
    get:
      tags:
        - leaseTemplates
      summary: Get paginated list of lease templates
      description: Returns all lease templates visible to the current user. Users with Admin or Manager roles can see all templates (PUBLIC and PRIVATE), while regular users can only see PUBLIC templates.
      parameters:
        - $ref: "#/components/parameters/PageIdentifier"
        - $ref: "#/components/parameters/MaxResults"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PaginatedResults"
                          - type: object
                            properties:
                              result:
                                type: array
                                items:
                                  $ref: "#/components/schemas/LeaseTemplate"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - leaseTemplates
      summary: Create new lease template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - requiresApproval
              properties:
                name:
                  $ref: "#/components/schemas/LeaseTemplateName"
                description:
                  $ref: "#/components/schemas/LeaseTemplateDescription"
                visibility:
                  $ref: "#/components/schemas/Visibility"
                maxSpend:
                  $ref: "#/components/schemas/MaxSpend"
                leaseDurationInHours:
                  $ref: "#/components/schemas/LeaseDurationInHours"
                budgetThresholds:
                  $ref: "#/components/schemas/BudgetThresholds"
                durationThresholds:
                  $ref: "#/components/schemas/DurationThresholds"
                requiresApproval:
                  type: boolean
                  example: true
                  description: Whether leases created from this template require approval
                costReportGroup:
                  $ref: "#/components/schemas/CostReportGroup"
                blueprintId:
                  allOf:
                    - $ref: "#/components/schemas/BlueprintUuid"
                    - nullable: true
                      description: Optional blueprint to attach to this lease template
                allowOwnerToShareLease:
                  type: boolean
                  description: Whether owners of leases created from this template can manage additional user assignments
                  example: false
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/LeaseTemplate"

        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /leaseTemplates/{leaseTemplateName}:
    parameters:
      - $ref: "#/components/parameters/LeaseTemplateIdPathParam"
    get:
      tags:
        - leaseTemplates
      summary: Get lease template by id
      description: Returns a specific lease template by ID. Users with Admin or Manager roles can access any template, while regular users can only access PUBLIC templates. Attempting to access a PRIVATE template as a regular user will return a 404 Resource Not Found.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/LeaseTemplate"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    put:
      tags:
        - leaseTemplates
      summary: Update lease template by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - requiresApproval
              properties:
                name:
                  type: string
                description:
                  type: string
                visibility:
                  $ref: "#/components/schemas/Visibility"
                requiresApproval:
                  type: boolean
                maxSpend:
                  $ref: "#/components/schemas/MaxSpend"
                leaseDurationInHours:
                  $ref: "#/components/schemas/LeaseDurationInHours"
                budgetThresholds:
                  $ref: "#/components/schemas/BudgetThresholds"
                durationThresholds:
                  $ref: "#/components/schemas/DurationThresholds"
                costReportGroup:
                  $ref: "#/components/schemas/CostReportGroup"
                blueprintId:
                  allOf:
                    - $ref: "#/components/schemas/BlueprintUuid"
                    - nullable: true
                      description: Optional blueprint to attach to this lease template
                allowOwnerToShareLease:
                  type: boolean
                  description: Whether owners of leases created from this template can manage additional user assignments
                  example: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/LeaseTemplate"

        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags:
        - leaseTemplates
      summary: Delete lease template by id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...